pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smart-nippo"
version = "0.1.0"
description = "日報入力支援ツール"
authors = [{name = "Your Name", email = "email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "typer>=0.9.0",
    "questionary>=2.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "pyperclip>=1.8.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pyright>=1.1.300",
    "ruff>=0.1.0",
    "pdoc>=14.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.scripts]
smart-nippo = "smart_nippo.cli.main:main"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "T20"]
ignore = []

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__"]
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.11"
pythonPlatform = "All"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=smart_nippo --cov-report=term-missing"